this is not even usable at the moment
i figured out how to make the id file of catalog and product visible
for node support and how to integrate them with HotChocolates
assumptions. But whenever I got the feeling to be getting it right this
time, HotChocolate pulls antoher id field called "iD" out its magic
hat and ruins my day. This is some task for future HaMster